cartoonify/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pillow>=11.1.0
pillow-heif>=1.3.0
python-multipart==0.0.12
